+++ .cbmc-batch/stubs/EVP_MD_CTX_free_no_pkey_stub.c
@@ -21,13 +21,6 @@
 #include <openssl/rsa.h>
 #include <proof_helpers/nondet.h>
 
-/* Abstraction of the EVP_MD_CTX struct */
-struct evp_md_ctx_st {
-    bool is_initialized;
-    EVP_PKEY *pkey;
-    size_t digest_size;
-};
-
 /*
  * Description: Cleans up digest context ctx and frees up the space allocated to it.
  * Use this stub when we are *certain* there is no pkey associated with the digest context.
